split file decbuf factory to own file and use a const for len bytes
